Move new engine version over from Safari project
There have been a lot of improvements and optimizations done in another
project (for a uni assignment) that have now been merged into the
original engine.

Some parts in the demo are broken(ish) as the engine's API has faced
some changes during the uni project, so those will have to be ported
over (or rewritten from scratch).